refactor (OmokBoardDbHelper): 데이터 베이스 관련 함수 헬퍼로 이동
